refactor example program + fix a bug and modify layoff API (and other small things)
