+++ examples/basic_rummy/ui/components/error.rs
@@ -1,0 +1,22 @@
+use ratatui::{
+    layout::{Alignment, Constraint, Direction, Layout, Rect},
+    style::{Color, Modifier, Style},
+    text::{Line, Span},
+    widgets::{Block, Borders, Paragraph, Wrap},
+    Frame,
+};
+use rummy::{cards::meld::Meldable, game::game::Game};
+use crate::app::{App, InputMode};
+
+pub fn render_error(f: &mut Frame, area: Rect, error_msg: &str) {
+    let content = vec![
+        Line::from(Span::styled("Error!", Style::default().add_modifier(Modifier::BOLD).fg(Color::Red))),
+        Line::from(""),
+        Line::from(error_msg),
+    ];
+
+    let paragraph = Paragraph::new(content)
+        .alignment(Alignment::Center)
+        .block(Block::default().borders(Borders::ALL).title("Error"));
+    f.render_widget(paragraph, area);
+}
